Добавляет презентор для списка точек
